+++ packages/indexer/src/core/index.ts
@@ -1,20 +1,17 @@
 import { BridgeContext } from '@nomad-xyz/sdk-bridge';
 import { ProcessorV2 } from './consumerV2';
 import { Orchestrator } from './orchestrator';
-import * as dotenv from 'dotenv';
 import { IndexerCollector } from './metrics';
 import { DB } from './db';
 import Logger from 'bunyan';
-import { run as runApi } from './api';
 import { getRedis } from './redis';
-dotenv.config({});
 
 export async function run(
   sdk: BridgeContext,
   db: DB,
   logger: Logger,
   metrics: IndexerCollector,
-) {
+): Promise<void> {
   const redis = getRedis();
 
   await redis.connect();
@@ -23,8 +20,6 @@ export async function run(
 
   const o = new Orchestrator(sdk, c, metrics, logger, db, redis);
   o.subscribeStatisticEvents();
-
-  if (process.env.DEBUG_PORT) runApi(o, logger.child({ span: 'debugApi' }));
 
   await o.init();
   await o.startConsuming();