+++ packages/indexer/src/core/sdk.ts
@@ -35,7 +35,10 @@ export async function getSdk(
   }
 
   sdk.domainNumbers.forEach((domain: number) => {
-    const name = sdk.mustGetDomain(domain).name.toUpperCase().replaceAll('-', '_');
+    const name = sdk
+      .mustGetDomain(domain)
+      .name.toUpperCase()
+      .replaceAll('-', '_');
     const rpcEnvKey = `${name}_RPC`;
     const defaultRPC = sdk.conf.rpcs[name]?.[0];
     const rpc = process.env[rpcEnvKey] || defaultRPC;