+++ packages/indexer/src/main.ts
@@ -1,22 +1,23 @@
 import * as core from './core';
-import * as api from './api';
 import { DB } from './core/db';
-import { BunyanLevel, createLogger } from './core/utils';
+import { run as apiRun } from './api';
+import { createLogger } from './core/utils';
 import { IndexerCollector } from './core/metrics';
 import { getSdk } from './core/sdk';
 import { startTokenUpdater } from './tokens';
-
-export type NomadEnvironment = 'development' | 'staging' | 'production';
-export type Program = 'api' | 'core';
-
-const environment = process.env.ENVIRONMENT! as NomadEnvironment;
-const configOverrideLocation = process.env.CONFIG_OVERRIDE_LOCATION;
-const program = process.env.PROGRAM! as Program;
-const logLevel = (process.env.LOG_LEVEL || 'debug') as BunyanLevel;
-const metricsPort = parseInt(process.env.METRICS_PORT || "9090");
+import {
+  program,
+  environment,
+  configOverrideLocation,
+  // metricsPort,
+  logLevel,
+  Program,
+} from './config';
 
 (async () => {
-  const logger = createLogger('indexer', environment, logLevel);
+  const loggerName = program === Program.API ? 'indexer_api' : 'indexer';
+
+  const logger = createLogger(loggerName, environment, logLevel);
   const m = new IndexerCollector(environment, logger);
 
   const sdk = await getSdk(configOverrideLocation || environment);
@@ -24,21 +25,15 @@ const metricsPort = parseInt(process.env.METRICS_PORT || "9090");
   const db = new DB(m, logger, sdk);
   await db.connect();
 
-  if (program === 'api') {
+  if (program === Program.API) {
     await startTokenUpdater(sdk, db, logger);
-    await api.run(db, logger);
+    await apiRun(db, logger);
     logger.info(`Finished api run`);
-  } else if (program === 'core') {
-    m.startServer(metricsPort);
+  } else if (program === Program.CORE) {
+    m.startServer(3000); // should be `metricsPort`, but thats a huge kludge now till next week :D
     await core.run(sdk, db, logger, m);
   } else {
-    logger.warn(`Starting all on the same process...`);
-    await startTokenUpdater(sdk, db, logger);
-    await Promise.all([
-      api.run(db, logger),
-      core.run(sdk, db, logger, m),
-    ]).catch((e) =>
-      logger.error(`Error happened during run of api or core:`, e),
-    );
+    logger.error(`Cannot run both at the same time`);
+    process.exit(1);
   }
 })();