+++ packages/contracts-core/CHANGELOG.md
@@ -2,5 +2,8 @@
 
 ### Unreleased
 
+### 1.1.0-rc.0
+
+- upgrade: contracts
 - feat: Export version from TS package
 - chore: disable metadata hash in hardhat compilation