+++ packages/indexer/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Unreleased
 
+- chore: refactor api service
+- chore: standardize script commands
+- feat: add config file
+- feat: track indexer api errors in sentry
 - feat: exposing api metrics on different port
 - feat: blocks to tip metric
 - feat: add version endpoint