+++ packages/multi-provider/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+### 1.0.0-rc.5
+
+- docs: update readme with link to documentation
 - update typescript eslint packages to fix lint script
 - remove unnecessary lint:fix script
 - standardize prettier